// secao-5/users-filter-project/src/app/interfaces/user/user.interface.ts
import { IAddress } from "./address.interface";
import { IStatus } from "./status.interface";

export interface IUser {
    nome: string
    email: string
    senha: string
    idade: string
    endereco: IAddress
    telefone: string
    ativo: boolean
    funcao: string
    dataCadastro: string
    status: IStatus
};